refactor: move isDblResourceReference and isProjectReference to platform-scripture-editor
These two type guards were only used by platform-scripture-editor, not internally
by platform-scripture. Moving them removes the @ts-ignore cross-extension
internal-path imports in the three consuming files.
