+++ extensions/src/platform-scripture-editor/src/resource-reference.utils.ts
@@ -1,0 +1,35 @@
+import type { DblResourceReference, ProjectReference } from 'platform-scripture';
+
+/**
+ * Checks if a {@link ResourceReference} is a {@link DblResourceReference}.
+ *
+ * @param item The resource reference to check
+ * @returns `true` if the item is a {@link DblResourceReference}, otherwise `false`
+ */
+export function isDblResourceReference(item: unknown): item is DblResourceReference {
+  return (
+    !!item &&
+    typeof item === 'object' &&
+    'type' in item &&
+    item.type === 'dblResource' &&
+    'id' in item &&
+    'name' in item
+  );
+}
+
+/**
+ * Checks if a {@link ResourceReference} is a {@link ProjectReference}.
+ *
+ * @param item The resource reference to check
+ * @returns `true` if the item is a {@link ProjectReference}, otherwise `false`
+ */
+export function isProjectReference(item: unknown): item is ProjectReference {
+  return (
+    !!item &&
+    typeof item === 'object' &&
+    'type' in item &&
+    item.type === 'project' &&
+    'id' in item &&
+    'name' in item
+  );
+}
