+++ extensions/src/platform-scripture/src/resource-reference-list.utils.ts
@@ -1,39 +1,4 @@
 import type { ProjectSettingValidator } from '@papi/core';
-import type { DblResourceReference, ProjectReference } from 'platform-scripture';
-
-/**
- * Checks if a {@link ResourceReference} is a {@link DblResourceReference}.
- *
- * @param item The resource reference to check
- * @returns `true` if the item is a {@link DblResourceReference}, otherwise `false`
- */
-export function isDblResourceReference(item: unknown): item is DblResourceReference {
-  return (
-    !!item &&
-    typeof item === 'object' &&
-    'type' in item &&
-    item.type === 'dblResource' &&
-    'id' in item &&
-    'name' in item
-  );
-}
-
-/**
- * Checks if a {@link ResourceReference} is a {@link ProjectReference}.
- *
- * @param item The resource reference to check
- * @returns `true` if the item is a {@link ProjectReference}, otherwise `false`
- */
-export function isProjectReference(item: unknown): item is ProjectReference {
-  return (
-    !!item &&
-    typeof item === 'object' &&
-    'type' in item &&
-    item.type === 'project' &&
-    'id' in item &&
-    'name' in item
-  );
-}
 
 export const isValidResourceReference = (item: unknown): boolean => {
   if (typeof item !== 'object' || !item) return false;
